fix: fix output dtype in test

+++ tests/test_drba.py
@@ -1,4 +1,5 @@
 import cv2
+import numpy as np
 import torch
 
 from ccvfi import AutoConfig, AutoModel, BaseConfig, ConfigType
@@ -29,7 +30,7 @@ class Test_DRBA:
             Outputs, _ = model.inference(inp, [-1, -0.5], [0], [0.5, 1], False, False, 1.0, None)
 
             for i in range(len(Outputs)):
-                out = Outputs[i].squeeze(0).permute(1, 2, 0).cpu().numpy() * 255
+                out = (Outputs[i].squeeze(0).permute(1, 2, 0).cpu().numpy() * 255).astype(np.uint8)
                 cv2.imwrite(str(ASSETS_PATH / f"test_out_{i}.jpg"), out)
 
                 assert calculate_image_similarity(eval_imgs[i], out)
